Bug fixes + dashboard enhancements + documentation updates
Data-layer fixes:
- HLP tracker: size_usd uses mark_px (not entry_px) for current market value
- Trade stream: correct liquidation side mapping (B→short, A→long)

Satellite audit fixes:
- co_occurrences: UNIQUE constraint + migration (INSERT OR IGNORE now deduplicates)
- Artemis seeder/pipeline: UTC-aware datetime for epoch computation
- Artemis profiler: collections.deque for O(1) FIFO popleft
- SHAP 0.49→0.50 (XGBoost 3.x base_score array compatibility)

Dashboard enhancements:
- data.html: candlestick chart (OHLC) replacing area chart, heatmap price lines overlay
- ml.html: prediction history panel (5th panel, newest-first table)
- dashboard.py: GET /api/ml/predictions/history endpoint

Documentation: updated 9 READMEs and context files across the project.

+++ data-layer/src/hynous_data/collectors/trade_stream.py
@@ -173,8 +173,8 @@ class TradeStream:
                     size_usd = abs(px * sz)
                     if size_usd >= 100:  # ignore dust liquidations
                         normalized_side = (
-                            "long" if side == "B"
-                            else "short" if side == "A"
+                            "short" if side == "B"
+                            else "long" if side == "A"
                             else side
                         )
                         users_list = trade.get("users", [])
